backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.122.0",
    "gitpython>=3.1.45",
    "pydantic>=2.12.5",
    "pywebview>=6.1",
    "strands-agents>=1.18.0",
    "uvicorn>=0.38.0",
]
